Fell behind in updates
Created updates regarding things I have researched and realized are not within my or other teammates' capabilities at this time. We could figure out a solution to it, but the time requirements are too great for a few guys working on a project in their free time. So, we've decided to no longer reinvent the wheel, but to create an application that is easy to maintain and will function as expected